backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Banias Baabe <banias@hotmail.de>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
notion-client = "^2.1.0"
langchain = "^0.0.346"
openai = "^1.3.7"
fastapi = {extras = ["uvicorn"], version = "^0.104.1"}
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-randomly = "*"
pytest-instafail = "*"
pytest-cov = "*"


[tool.poetry.group.linters]
optional = true

[tool.poetry.group.linters.dependencies]
mypy = "*"
interrogate = "*"
ruff = "*"
bandit = {extras = ["toml"], version = "*"}

[tool.poetry.group.fixers]
optional = true

[tool.poetry.group.fixers.dependencies]
docformatter = "*"
codespell = "^2.2.6"

[tool.ruff]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.interrogate]
ignore-init-method = true
fail-under = 60
exclude = ["frontend","backend/.venv","backend/tests"]

[tool.mypy]
disable_error_code = ["import-untyped","call-overload"]

[tool.bandit]
exclude_dirs = ["venv","tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
